// src/app/shared/services/camunda-api/model/processDefinitionDto.ts
//
//   Copyright 2026 Deutsches Zentrum für Luft- und Raumfahrt e.V.
//
//   Licensed under the Apache License, Version 2.0 (the "License");
//   you may not use this file except in compliance with the License.
//   You may obtain a copy of the License at
//
//       http://www.apache.org/licenses/LICENSE-2.0
//
//   Unless required by applicable law or agreed to in writing, software
//   distributed under the License is distributed on an "AS IS" BASIS,
//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
//   See the License for the specific language governing permissions and
//   limitations under the License.
//

/**
 * Camunda Platform REST API
 * OpenApi Spec for Camunda Platform REST API.
 *
 * The version of the OpenAPI document: 7.16.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface ProcessDefinitionDto {
  /**
   * The id of the process definition
   */
  id: string;
  /**
   * The key of the process definition, i.e., the id of the BPMN 2.0 XML process definition.
   */
  key: string;
  /**
   * The category of the process definition.
   */
  category?: string | null;
  /**
   * The description of the process definition.
   */
  description?: string | null;
  /**
   * The name of the process definition.
   */
  name?: string | null;
  /**
   * The version of the process definition that the engine assigned to it.
   */
  version: number;
  /**
   * The file name of the process definition.
   */
  resource?: string | null;
  /**
   * The deployment id of the process definition.
   */
  deploymentId?: string | null;
  /**
   * The file name of the process definition diagram, if it exists.
   */
  diagram?: string | null;
  /**
   * A flag indicating whether the definition is suspended or not.
   */
  suspended?: boolean | null;
  /**
   * The tenant id of the process definition.
   */
  tenantId?: string | null;
  /**
   * The version tag of the process definition.
   */
  versionTag?: string | null;
  /**
   * History time to live value of the process definition. Is used within [History cleanup](https://docs.camunda.org/manual/7.16/user-guide/process-engine/history/#history-cleanup).
   */
  historyTimeToLive?: number | null;
  /**
   * A flag indicating whether the process definition is startable in Tasklist or not.
   */
  startableInTasklist?: boolean | null;
}